CEDS-3620 Make email package and correct nits
