+++ app/models/email/UndeliverableInformation.scala
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package models.email
+
+import org.joda.time.DateTime
+
+import play.api.libs.json.{Json, OFormat}
+
+final case class UndeliverableInformation(
+  subject: String,
+  eventId: String,
+  groupId: String,
+  timestamp: DateTime,
+  event: UndeliverableInformationEvent
+)
+
+object UndeliverableInformation {
+  import play.api.libs.json.JodaReads._
+  import play.api.libs.json.JodaWrites._
+  implicit val format: OFormat[UndeliverableInformation] = Json.format[UndeliverableInformation]
+}
